refactor: move constants to const js
